redis + nodebb
Settings and the like go to /data.  Somewhat of a WIP, wanted to get feedback from the community on it.  

Thanks!

FYI, the next thing I am doing with this is adding config by environment variables.  I recommended discourse for a long time, but reality is reality:  Ruby isn't an appropriate language for today's internet.  (though I'll be frank and say that I prefer go to node (but don't program that much anyway, so maybe my opinion is rather invalid-- I am the implementer of others' code (hence the ruby-hate), usually. 

Also, I just had the realization that I'd gone about this wrong, and will be submitting two dockerfiles later today that will separately implement redis and nodebb.  I'll use the microbox redis instance, since it's 7MB.